+++ src/routes/transactions/entities/transaction-info.entity.ts
@@ -1,21 +1,4 @@
-import {
-  ApiExtraModels,
-  ApiProperty,
-  ApiPropertyOptional,
-  type ApiPropertyOptions,
-  getSchemaPath,
-} from '@nestjs/swagger';
-
-import { CreationTransactionInfo } from '@/routes/transactions/entities/creation-transaction-info.entity';
-import { CustomTransactionInfo } from '@/routes/transactions/entities/custom-transaction.entity';
-import { SettingsChangeTransaction } from '@/routes/transactions/entities/settings-change-transaction.entity';
-import { TransferTransactionInfo } from '@/routes/transactions/entities/transfer-transaction-info.entity';
-import { SwapOrderTransactionInfo } from '@/routes/transactions/entities/swaps/swap-order-info.entity';
-import { SwapTransferTransactionInfo } from '@/routes/transactions/swap-transfer-transaction-info.entity';
-import { TwapOrderTransactionInfo } from '@/routes/transactions/entities/swaps/twap-order-info.entity';
-import { NativeStakingDepositTransactionInfo } from '@/routes/transactions/entities/staking/native-staking-deposit-info.entity';
-import { NativeStakingValidatorsExitTransactionInfo } from '@/routes/transactions/entities/staking/native-staking-validators-exit-info.entity';
-import { NativeStakingWithdrawTransactionInfo } from '@/routes/transactions/entities/staking/native-staking-withdraw-info.entity';
+import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
 
 export enum TransactionInfoType {
   Creation = 'Creation',
@@ -30,18 +13,6 @@ export enum TransactionInfoType {
   NativeStakingWithdraw = 'NativeStakingWithdraw',
 }
 
-@ApiExtraModels(
-  CreationTransactionInfo,
-  CustomTransactionInfo,
-  SettingsChangeTransaction,
-  TransferTransactionInfo,
-  SwapOrderTransactionInfo,
-  SwapTransferTransactionInfo,
-  TwapOrderTransactionInfo,
-  NativeStakingDepositTransactionInfo,
-  NativeStakingValidatorsExitTransactionInfo,
-  NativeStakingWithdrawTransactionInfo,
-)
 export class TransactionInfo {
   @ApiProperty({ enum: TransactionInfoType })
   type: TransactionInfoType;
@@ -55,22 +26,4 @@ export class TransactionInfo {
     this.type = type;
     this.humanDescription = humanDescription;
   }
-}
-
-export function getTxInfoSchema(): ApiPropertyOptions {
-  return {
-    nullable: true,
-    oneOf: [
-      { $ref: getSchemaPath(CreationTransactionInfo) },
-      { $ref: getSchemaPath(CustomTransactionInfo) },
-      { $ref: getSchemaPath(SettingsChangeTransaction) },
-      { $ref: getSchemaPath(TransferTransactionInfo) },
-      { $ref: getSchemaPath(SwapOrderTransactionInfo) },
-      { $ref: getSchemaPath(SwapTransferTransactionInfo) },
-      { $ref: getSchemaPath(TwapOrderTransactionInfo) },
-      { $ref: getSchemaPath(NativeStakingDepositTransactionInfo) },
-      { $ref: getSchemaPath(NativeStakingValidatorsExitTransactionInfo) },
-      { $ref: getSchemaPath(NativeStakingWithdrawTransactionInfo) },
-    ],
-  };
 }
