+++ src/routes/transactions/entities/transaction.entity.ts
@@ -9,7 +9,8 @@ import { ModuleExecutionInfo } from '@/routes/transactions/entities/module-execu
 import { MultisigExecutionInfo } from '@/routes/transactions/entities/multisig-execution-info.entity';
 import { SafeAppInfo } from '@/routes/transactions/entities/safe-app-info.entity';
 import { TransactionStatus } from '@/routes/transactions/entities/transaction-status.entity';
-import { getTxInfoSchema, TransactionInfo } from './transaction-info.entity';
+import { TransactionInfo } from './transaction-info.entity';
+import { getTxInfoSchema } from '../helpers/tx-info-schema';
 
 @ApiExtraModels(TransactionInfo)
 export class Transaction {
